fix(onboard): keep onboard.ts net-neutral + thread WSL flag downstream
