+++ src/lib/onboard/docker-gpu-patch.test.ts
@@ -105,40 +105,6 @@ describe("docker-gpu-patch", () => {
     ).toBe(false);
   });
 
-  it("ignores NEMOCLAW_DOCKER_GPU_PATCH=0 on Docker Desktop WSL where the patch is required", () => {
-    const logs: string[] = [];
-    expect(
-      shouldApplyDockerGpuPatch(
-        { sandboxGpuEnabled: true },
-        {
-          env: { NEMOCLAW_DOCKER_GPU_PATCH: "0" },
-          platform: "linux",
-          dockerDriverGateway: true,
-          dockerDesktopWsl: true,
-          log: (message) => {
-            logs.push(message);
-          },
-        },
-      ),
-    ).toBe(true);
-    expect(logs.some((line) => /NEMOCLAW_DOCKER_GPU_PATCH=0 ignored/i.test(line))).toBe(true);
-    expect(logs.some((line) => /NEMOCLAW_SANDBOX_GPU=0/.test(line))).toBe(true);
-  });
-
-  it("still honors NEMOCLAW_DOCKER_GPU_PATCH=0 when not on Docker Desktop WSL", () => {
-    expect(
-      shouldApplyDockerGpuPatch(
-        { sandboxGpuEnabled: true },
-        {
-          env: { NEMOCLAW_DOCKER_GPU_PATCH: "0" },
-          platform: "linux",
-          dockerDriverGateway: true,
-          dockerDesktopWsl: false,
-        },
-      ),
-    ).toBe(false);
-  });
-
   it("builds clone args that preserve OpenShell labels and runtime settings", () => {
     const args = buildDockerGpuCloneRunArgs(inspectFixture(), buildDockerGpuMode("gpus"));
